requirements-frontend.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
python-dotenv
google-cloud-alloydb-connector[pg8000]
pg8000
